ark_scale

Module hazmat

Hazardous projective serialization and deserialization

Do NOT use this module in wire formats or other untrusted cases.

Aside from not being cannonical, we caution that projective coordinates actually leak secret key material, which makes them extraordinarily dangerous.

We do not implement curve or subgroup check for elliptic curve points in projective coordinates, nor do we validate their base field elements. You diserve to be p0wned if you use them in an untrusted enviroment.

Modules

sw

te

Structs

ArkScaleProjective

Functions

ark_sw_decode_from

ark_sw_encode_to

ark_te_decode_from

ark_te_encode_to

scale_sw_decode_from

scale_sw_encode_to

scale_te_decode_from

scale_te_encode_to