ark_scale

Trait ArkScaleMaxEncodedLen

pub trait ArkScaleMaxEncodedLen {
    // Required method
    fn max_encoded_len(compress: Compress) -> usize;
}

An orphan rules helper which provides impl scale::MaxEncodedLen for ArkScale<T,WIRE>

We suggest wrapper types in preference to this trait becuase on-chain verifier code becomes substrate aware anyways whenever you choose curves, via a conditional ark-substrate dependence. Also wrappers can improve documentation and errors.

Required Methods

fn max_encoded_len(compress: Compress) -> usize

Upper bound, in bytes, of the maximum encoded size of this item.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.